+++ libs/public-health/src/vivarium/public_health/treatment/magic_wand.py
@@ -96,6 +96,7 @@ class AbsoluteShift(Component):
             f"{self.target.name}.{self.target.measure}",
             modifier=self.intervention_effect,
             required_resources=["age"],
+            description="Replace the target with the configured intervention value",
         )
 
     ##################################